fix(apigateway): permission error in lambda integration when function name is modified (#8813)
Changing function name triggers a resource replacement - the old
function is removed and replaced with a new function.
However, the RestApi deployment remains untouched and is still pointing
at the ARN of the, now stale, function.

The fix is to trigger a new deployment if the function name changes.

fixes #5306


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*