chore(apigateway): modernize Integration.bind() (#8814)
Currently, the Method construct reaches into the internals of the
Integration class to construct itself, via the _props() internal method.

Change to a more recent pattern where the bind() returns a result that
contains all of the information that the Method class requires to
finalize the bind.

Motivation for the change
This change - https://github.com/aws/aws-cdk/pull/8813 - requires a
property returned to the Method that is not user specified. The change
already introduces this pattern. This PR takes it a bit further and
applies the same logic to existing properties, so that there is now only
one way for the Method to get the result of the bind.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*