chore: remove TSLint
On every build, we are basically compiling our code 3 times: once with
jsii, once with eslint and once with tslint. Also, you need 2 plugins
installed into your IDE to get linter erroring and fixing as you type.

Between these, TSLint is deprecated, so why are we even doing this
anymore? Reduce build times by removing TSLint and migrating all the
rules we were getting from it to ESLint.

Make all packages lint correctly, which was also not the case.