fix(ec2): VpcEndpoint AZ lookup fails for AWS services (#8386)
fix(ec2): VpcEndpoint AZ lookup fails for AWS services
----

This closes https://github.com/aws/aws-cdk/issues/8372

InterfaceVpcEndpointAwsService objects currently use the `AWS.Region` token to define the service name. This results in an unresolved string, which cannot be resolved until CloudFormation deployment time.

When using the `lookupSupportedAzs` functionality, a ContextProvider makes AWS calls (DescribeVpcEndpointServices) in which it passes the service name, to look up the AZs. ContextProviders cannot be given tokens. Hence, passing in an InterfaceVpcEndpointAwsService results in a synthesis-time failure.

This ticket switches InterfaceVpcEndpointAwsService from using AWS.Region to a Lazy string which resolves to the stack's region. For region-agnostic stacks, the behavior is unchanged. For stacks with a specified region, the region is resolved.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*