chore: make tests independent of asset hashes
Introduce a `canonicalizeTemplate` function in `@aws-cdk/assert` which
translates templates to a common form w.r.t. asset hashes.

Use this in some tests and in `cdk-integ-assert` to make the tests
succeed if all that is different is the specific asset hash.

Currently only supports legacy assets, should be updated for new-style
assets when we switch to those.

This change is necessary to unblock other build improvements/changes
such as #8946.