chore: jest configs should not rely on source tree layout (#8970)
The way jest configs are currently inheriting from a parent
jest config makes it hard to do an out-of-source build, which we
want for future build improvements.

They used to go `require('../../../tools/cdk-build-tools/...')`,
relying on an exact layout in the source tree.

Given that `cdk-build-tools` is just a dependency of every
package, doing `require('cdk-build-tools/...')` works just as well
and only relies on the standard NPM contract (`node_modules` having
correct symlinks) and not source tree layout.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*