chore: make matchers written for `pipelines` available globally
There are useful matchers like `stringLike` and `encodedJson` which
are useful in many many places. Push them to a central location.

Also in this commit: turns out that the tests in the
`@aws-cdk/pipelines` package weren't actually being run, because
we didn't have:

```
"cdk-build": {
  "jest": true
}
```

set in `package.json`... o_O. Correct that oversight as well,
now actually run the tests.