chore: remove TSLint (#8946)
On every build, we are basically compiling our code 3 times: once with
jsii, once with eslint and once with tslint. 

Between these, TSLint is deprecated, so why are we even doing this
anymore? Reduce build times by removing TSLint and migrating all the
rules we were getting from it to ESLint.

Also, with the current setup you need 2 plugins installed into your
IDE to get all linting errors reported as-you-type. Reduce to one
plugin.

This change makes `lerna run lint` run successfully, which was also not the case.

This change shaves 5 minutes (25%) off of the PR build (21 minutes down to 16 minutes).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*