chore: move `jest-junit` dependency to top-level
`"jest-junit"` is a new dependency for all packages, used as a Jest reporter
in the default Jest config located in `cdk-build-tools`.

`jest-junit` was added as a dependency to `cdk-build-tools`, but when
resolving reporters Jest starts resolving from the *package under test*.

That means that `jest-junit` should be a `devDependency` for every
individual package; or more easily: add the dependency to the root of
the repository so that it is automatically in scope for every package
inside it.

It used to accidentally work before this change, because the dependency
hoisting performed by Yarn moved the dependency up to the root of the
tree where it would be found by the jest run.