Revert "chore(core): allow the bundler to re-use pre-existing bundler output (#8916)" (#9037)
This logic to reuse bundler output is faulty. Once an asset is bundled, the bundle directory is moved into the cloud assembly (cdk.out) and therefore the logic that attempts to reuse the bundle directory only materializes in the case where moveSync decides to copy the directory and not just move it.

This reverts commit 31d6e6596cde03daa7de48aab8aaae1277fd405e.