fix(stepfunctions): Map state does not render JSON paths from state input (#9008)
The Map state property parameters support JSON path, but
the CDK does not render these as it does with other properties
that support using JSON path.

This change renders the properties so that Json path strings
which start with $. have their keys adjusted to reference the
state input field they reference.

Closes #8992 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*