fix(cli): diff against multiple stacks do not always fail if any have a diff (#7690)
### Commit Message
Currently, the diff command only remembers if there are any diffs
with the last stack that was diffed. So, if all of your stacks have
a diff except for the last one, the diff command would exit
successfully even if told to fail on diff.

This fix corrects the behavior so that if any of the stacks have
a diff, then the exit code would be non-zero if configured to do so.

fixes #7492
### End Commit Message

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*