feat(cloudfront): Initial CloudFront redesign (#8982)
See https://github.com/aws/aws-cdk-rfcs/pull/174 for the design that inspired this implementation.

The goal of this initial PR is to get the new Distribution construct to the point where it supports the most common use cases, in terms of single- and multi-origin and behaviors with the most common properties exposed and working. This is not intended to be a complete/exhaustive implementation, but does allow for users with S3 buckets to create CloudFront distributions more easily than the existing construct today.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*