feat(aws-stepfunctions-tasks): allow lambda invocations to combine input and function results (#9022)
Step Functions allows a third way to invoke Lambda functions that
was previously unimplemented in LambdaInvoke by specifying the
function ARN as the Resource. As implemented, LambdaInvoke does
not allow the combination of the inputPath and the Lambda function
result without including extra metadata. A suggestion workaround was
to add a `Pass` state to manipulate the output. This change implements
that option without requiring an extra state transition.

fixes #8943


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*