fix(cli): cli integ tests do not have a unique stack prefix (#9165)
Stack prefixes only included the codebuild project _name_ and not the build id. There was also some legacy support in which the test prefix was `cdk-toolkit-canary` or `cdk-toolkit-integration` and it seems like these prefixes are no longer required.

This resulted in canaries failing when two instances executed concurrently: the "cleanup" code of one canary deleted stacks in-flight for the other canary.

We now simply use a timestamp to generate a unique prefix for each test run in order to isolate these stacks.