fix(lambda-nodejs): permission denied on npm cache (#9167)
The npm cache was owned by the user 1000 so if the image was run by
a user other than 1000 installing via npm (when using the `nodeModules`
prop) resulted in a `EACCES` error on the cache folder. Yarn automatically
falls back to `/tmp` in this case.

Move npm cache folder creation to the end of the Dockerfile so that no
root content has been created in it when the image starts. Set cache
folder permission to 777.

Also disable npm update notification and add tests for the Dockerfile.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*