feat(lambda): higher level construct for python
This construct is heavily inspired by @jogold's Nodejs construct
introduced in #5532. The Python equivalent is simpler in that no
external bundler is necessary; it's also possible that this construct
doesn't relieve enough pain to be introduced and maintained. I'd
personally love to remove some of the boilerplate needed to pip
install my python dependencies and enjoy the symmetry of defining
`lambda.PythonFunction` alongside `lambda.NodejsFunction`.