fix(appsync): resolver unable to set pipelineConfig (#9093)
**[ISSUE]**
`pipelineConfig` was be labeled as `undefined` while using the `Resolver` class instead of `createResolver`. Also, no way to set `kind` parameter for resolvers, so implemented that as well.

**[APPROACH]**
Created a property that takes `pipelineConfig` for `AppSync` functions. 

**[NOTE]**
`pipelineConfig` takes a string array for the name of `AppSync Functions` not `Lambda Functions`

Fixes #6923  

BREAKING CHANGE: `pipelineConfig` is now an array of `string` instead of `CfnResolver.PipelineConfigProperty` for usability.
- **appsync**: `pipelineConfig` parameter takes in `string []`

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*