fix(certificatemanager): Fix DNS validation for wildcard certificates
If a certificate with automatic (Route53) DNS validation contains both a base
domain name and the wildcard for that domain (e.g., `example.com` and
`*.example.com`), the corresponding DNS validation records are identical.
This seems to have caused problems for the automated CloudFormation DNS
validation.

Solving the problem by removing the redundant wildcard entries from the
DomainValidationOption.

fixes #9248