chore(cloudfront): small refactoring of the Origin API (#9281)
Change the Origin to an interface, from an abstract class, and change its `bind` protocol to return an `OriginBindConfig` interface.

This is in preparation for handling Origin Groups in #9109 - when time comes to handle Origin Groups, we will add a new (optional) property to `OriginBindConfig`, of type `CfnDistribution.OriginGroupProperty`, and handle it in `Distribution`.

BREAKING CHANGE: the property Origin.domainName has been removed

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*