chore: remove an additional "prepare" from apigateway
Follow up on #9410 and remove an additional usage of `prepare` in the API Gateway library. In this case we leveraged `prepare()` to record dependencies between all Deployment resources and all Method resources in the APIGW.

The solution is to perform two-sided bookkeeping while methods/deployments are defined and record their dependencies in-band.

We also refactored the way the code in `LatestDeployment` to be slightly more readable.