feat(eks): default masters role
If `mastersRole` is not specified, we now define a default IAM role that can be assumed by anyone (with permissions) in the account.
    
If `mastersRole` is not specified, we now define a default IAM role that can be assumed by anyone (with permissions) in the account.
    
This will allow users to interact with the cluster through `kubectl` by issuing the `aws eks update-kubeconfig` command with the appropriate `--role-arn` option, as specified in the CFN output.
    
Fixes #9463