feat(eks): default masters role (#9464)
If `mastersRole` is not specified, we now define a default IAM role that can be assumed by anyone (with permissions) in the account.
        
This will allow users to interact with the cluster through `kubectl` by issuing the `aws eks update-kubeconfig` command with the appropriate `--role-arn` option, as specified in the CFN output.
    
Fixes #9463


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*