fix(cfn-include): no longer concatenate elements of Fn::Join without tokens
Fn.join() has some logic inside of it to simplify the expressions to concatenate the
array elements that do not contain any Tokens inside of them.
We don't want to do it in cfn-include though,
as that causes a diff from the original template.
So, wrap the array given as the second argument to Fn::Join into a Token,
to prevent the concatenation logic from triggering.