feat(core): deprecate "Construct.node" in favor of "Construct.construct"
To reduce the chance for conflicts with generated domain-specific properties in subclasses of `Construct` (see [terraform-cdk issue]), we decided that we will rename `node` to `construct`, which is less prevalent.

We plan to remove the `node` API in the next major version of the AWS CDK, and therefore recommend users to migrate their code to use `construct` instead.

Part of https://github.com/aws/aws-cdk-rfcs/issues/192
[terraform-cdk issue]: https://github.com/hashicorp/terraform-cdk/pull/230