fix(pipelines): manual approval of changeset uses wrong ordering (#9508)
From `addApplication` or `addStackArtifactDeployment` method.

Currently, when calling `addStackArtifactDeployment`, you can pass in `AddStackOptions`, but the `executeRunOrder` property is not respected later in the process where `DeployCdkStackAction.fromStackArtifact` is called.
From addApplication, this property is used when manualApprovals has been opted in to, but as it is not respected, it results in bug #9101

Fixes #9101 (I think!)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*