feat(core): deprecate "Construct.node" in favor of "Construct.construct" (#9557)
To reduce the chance for conflicts with generated domain-specific properties in subclasses of `Construct` (see [terraform-cdk issue]), we decided that we will rename `node` to `construct`, which is less prevalent.

We plan to remove the `node` API in the next major version of the AWS CDK, and therefore recommend users to migrate their code to use `construct` instead.

See discussion in the [RFC].

[terraform-cdk issue]: https://github.com/hashicorp/terraform-cdk/pull/230
[RFC]: https://github.com/aws/aws-cdk-rfcs/blob/master/text/0192-remove-constructs-compat.md#10-construct-node

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*