chore: Mergify rule requiring two (or more) approvals on certain PRs
For security-related or other broadly-impacting PRs, having multiple reviewers
can be beneficial. Today, PRs can be manually tagged with 'pr/do-not-merge' and
untagged after an artibrary point. An explicit label
('pr/requires-two-approvers`) makes this more explicit and allows for up-front
tagging by the PR author.

Feedback on the idea welcome.