feat: "stack relative exports" flag
Previously, CloudFormation `Exports` generated by cross-stack references
were named after a static part of the construct tree (everything after
the 2nd component). This must have been left over from the days
when the construct tree structure was strictly

```
App -> Stack -> Other constructs...
```

And so it would skip the `App` and `Stack` levels, leaving a construct
path that was relative to the containing stack.

For a while, `Stacks` have been able to occur at multiple levels, but
this code was never updated. The result is that this now produces
considerable breakage if stacks are moved around (motivating example:
moving `Stack`s into `Stages` for the benefit of deploying them
using CDK Pipelines).

This behavior cannot be fixed without breaking most CDK applications
that currently exist, so it is being introduced as a feature flag,
which should be enabled on new applications and made default in V2,
but for now is opt-in and to be enabled by people migrating their
existing apps onto CDK Pipelines.

The new behavior is to explicitly build the export name relative to the
containing Stack, as opposed to relative to construct tree level 2.