chore: revert introduction of "Construct.construct" member (#9637)
The addition of the new `construct` member leads to problems in the C# code generation, where it would properly be called `Construct.Construct`: a member may not have the same name as the class, because that is the name of the class constructor (see https://github.com/aws/jsii/issues/1880).

Our build is currently broken because of this. Revert the renames to unblock the build, giving us the opportunity to tackle this problem afresh next week.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*