feat(codebuild): add support for GPU build images (#8879)
CodeBuild has added support for running builds on machines with GPU drivers.
This required some changes to the protocol between `IBuildImage` and `Project`,
as those images are hosted in a public ECR repository that the image must grant the Project's Role access to.
Introduced a `bind()` method to `IBuildImage` -
to maintain backwards comaptibility,
made it optional.

Fixes #8408 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*