feat(autoscaling): CloudFormation init for ASGs
Introduces CloudFormation init (cfn-init) support for autoscaling groups.
This builds on the previous work (#9065) and (#9664) that introduced init
support for instances.

This change also reworks the existing signaling functionality, as this becomes
even more important with cfn-init.

A final change was to adjust the (internal) interface of the `_attach` method to
accept an `ec2.OperatingSystemType` instead of the internal-only
`ec2.InitPlatform` type which wasn't exported and unavailable within the asg
module. This change is also in the #9664 diff, as both are pending.

Credit for 90% of this goes to @rix0rrr; all cfn-init support was pair-
programmed, but the ASG stuff was much more heavily done by him.